+++ commands/systems/http_delete.py
@@ -63,10 +63,8 @@ class CommandHandler(CommandHandlerBase):
 
     @classmethod
     def display_results(self, redfishConfig):
-        Trace.log(TraceLevel.INFO, '')
-        Trace.log(TraceLevel.INFO, '[] URL        : {}'.format(self.startingurl))
-        Trace.log(TraceLevel.INFO, '[] Status     : {}'.format(self.link.urlStatus))
-        Trace.log(TraceLevel.INFO, '[] Reason     : {}'.format(self.link.urlReason))
+        if self.link != None:
+            self.link.print_status()
 
         if (self.link != None and self.link.jsonData != None):
             Trace.log(TraceLevel.INFO, '[] HTTP Data  : {}'.format(self.link.jsonData))