+++ commands/systems/show_fans.py
@@ -111,10 +111,7 @@ class CommandHandler(CommandHandlerBase):
     def display_results(self, redfishConfig):
         if (self.link != None):
             if (self.link.valid == False):
-                print('')
-                print(' [] URL        : {}'.format(self.link.url))
-                print(' [] Status     : {}'.format(self.link.urlStatus))
-                print(' [] Reason     : {}'.format(self.link.urlReason))
+                self.link.print_status()
     
             else:
                 print('')