+++ commands/systems/show_sessions.py
@@ -130,10 +130,7 @@ class CommandHandler(CommandHandlerBase):
     def display_results(self, redfishConfig):
         # self.print_banner(self)
         if (self.link.valid == False):
-            print('')
-            print(' [] URL        : {}'.format(self.link.url))
-            print(' [] Status     : {}'.format(self.link.urlStatus))
-            print(' [] Reason     : {}'.format(self.link.urlReason))
+            self.link.print_status()
 
         else:
             # Sort the list