+++ commands/systems/show_thermal.py
@@ -165,11 +165,8 @@ class CommandHandler(CommandHandlerBase):
     def display_results(self, redfishConfig):
         # self.print_banner(self)
         if (self.link != None):
-            if (self.link == None or self.link.valid == False):
-                print('')
-                print(' [] URL        : {}'.format(self.link.url))
-                print(' [] Status     : {}'.format(self.link.urlStatus))
-                print(' [] Reason     : {}'.format(self.link.urlReason))
+            if (self.link.valid == False):
+                self.link.print_status()
     
             else:
                 print('')